+++ sc2reader/events/game.py
@@ -13,8 +13,6 @@ class GameEvent(Event):
     """
     This is the base class for all game events. The attributes below are universally available.
     """
-    name = 'GameEvent'
-
     def __init__(self, frame, pid):
         #: The id of the player generating the event. This is 16 for global non-player events.
         #: Prior to Heart of the Swarm this was the player id. Since HotS it is
@@ -34,6 +32,9 @@ class GameEvent(Event):
         #: A flag indicating if it is a local or global event.
         self.is_local = (pid != 16)
 
+        #: Short cut string for event class name
+        self.name = self.__class__.__name__
+
     def _str_prefix(self):
         player_name = self.player.name if getattr(self, 'pid', 16) != 16 else "Global"
         return "%s\t%-15s " % (Length(seconds=int(self.frame/16)), player_name)
@@ -47,9 +48,6 @@ class GameStartEvent(GameEvent):
     Recorded when the game starts and the frames start to roll. This is a global non-player
     event.
     """
-
-    name = 'GameStartEvent'
-
     def __init__(self, frame, pid, data):
         super(GameStartEvent, self).__init__(frame, pid)
 
@@ -58,9 +56,6 @@ class PlayerLeaveEvent(GameEvent):
     """
     Recorded when a player leaves the game.
     """
-
-    name = 'PlayerLeaveEvent'
-
     def __init__(self, frame, pid, data):
         super(PlayerLeaveEvent, self).__init__(frame, pid)
 
@@ -70,9 +65,6 @@ class UserOptionsEvent(GameEvent):
     This event is recorded for each player at the very beginning of the game before the
     :class:`GameStartEvent`.
     """
-
-    name = 'UserOptionsEvent'
-
     def __init__(self, frame, pid, data):
         super(UserOptionsEvent, self).__init__(frame, pid)
         #:
@@ -118,12 +110,8 @@ def create_command_event(frame, pid, data):
         return SelfAbilityEvent(frame, pid, data)
 
 
-class PlayerActionEvent(GameEvent):
-    name = 'PlayerActionEvent'
-
-
 @loggable
-class AbilityEvent(PlayerActionEvent):
+class AbilityEvent(GameEvent):
     """
     Ability events are generated when ever a player in the game issues a command
     to a unit or group of units. They are split into three subclasses of ability,
@@ -133,11 +121,6 @@ class AbilityEvent(PlayerActionEvent):
     See :class:`LocationAbilityEvent`, :class:`TargetAbilityEvent`, and :class:`SelfAbilityEvent`
     for individual details.
     """
-
-    name = 'AbilityEvent'
-
-    is_player_action = True
-
     def __init__(self, frame, pid, data):
         super(AbilityEvent, self).__init__(frame, pid)
 
@@ -236,9 +219,6 @@ class LocationAbilityEvent(AbilityEvent):
     Note that like all AbilityEvents, the event will be recorded regardless
     of whether or not the command was successful.
     """
-
-    name = 'LocationAbilityEvent'
-
     def __init__(self, frame, pid, data):
         super(LocationAbilityEvent, self).__init__(frame, pid, data)
 
@@ -265,9 +245,6 @@ class TargetAbilityEvent(AbilityEvent):
 
     Note that all AbilityEvents are recorded regardless of whether or not the command was successful.
     """
-
-    name = 'TargetAbilityEvent'
-
     def __init__(self, frame, pid, data):
         super(TargetAbilityEvent, self).__init__(frame, pid, data)
 
@@ -318,9 +295,6 @@ class SelfAbilityEvent(AbilityEvent):
 
     Note that all AbilityEvents are recorded regardless of whether or not the command was successful.
     """
-
-    name = 'SelfAbilityEvent'
-
     def __init__(self, frame, pid, data):
         super(SelfAbilityEvent, self).__init__(frame, pid, data)
 
@@ -329,7 +303,7 @@ class SelfAbilityEvent(AbilityEvent):
 
 
 @loggable
-class SelectionEvent(PlayerActionEvent):
+class SelectionEvent(GameEvent):
     """
     Selection events are generated when ever the active selection of the
     player is updated. Unlike other game events, these events can also be
@@ -340,10 +314,6 @@ class SelectionEvent(PlayerActionEvent):
     by non-player actions. When a player action updates a control group
     a :class:`HotkeyEvent` is generated.
     """
-
-    name = 'SelectionEvent'
-    is_player_action = True
-
     def __init__(self, frame, pid, data):
         super(SelectionEvent, self).__init__(frame, pid)
 
@@ -392,39 +362,32 @@ class SelectionEvent(PlayerActionEvent):
 def create_control_group_event(frame, pid, data):
     update_type = data['control_group_update']
     if update_type == 0:
-        return SetToHotkeyEvent(frame, pid, data)
+        return SetControlGroupEvent(frame, pid, data)
     elif update_type == 1:
-        return AddToHotkeyEvent(frame, pid, data)
+        return AddToControlGroupEvent(frame, pid, data)
     elif update_type == 2:
-        return GetFromHotkeyEvent(frame, pid, data)
+        return GetControlGroupEvent(frame, pid, data)
     elif update_type == 3:
         # TODO: What could this be?!?
-        return HotkeyEvent(frame, pid, data)
+        return ControlGroupEvent(frame, pid, data)
 
 
 @loggable
-class HotkeyEvent(PlayerActionEvent):
+class ControlGroupEvent(GameEvent):
     """
-    Hotkey events are recorded when ever a player action modifies a control
-    group. I know that calling control group events hotkey events doesn't make
-    sense but for backwards compatibility I haven't changed it yet. Sorry.
-
-    There are three kinds of hotkey events, generated by each of the possible
+    ControlGroup events are recorded when ever a player action modifies or accesses a control
+    group. There are three kinds of events, generated by each of the possible
     player actions:
 
-    * :class:`SetToHotkeyEvent` - Recorded when a user sets a control group (ctrl+#).
-    * :class:`GetFromHotkeyEvent` - Recorded when a user retrieves a control group (#).
-    * :class:`AddToHotkeyEvent` - Recorded when a user adds to a control group (shift+ctrl+#)
+    * :class:`SetControlGroup` - Recorded when a user sets a control group (ctrl+#).
+    * :class:`GetControlGroup` - Recorded when a user retrieves a control group (#).
+    * :class:`AddToControlGroup` - Recorded when a user adds to a control group (shift+ctrl+#)
 
     All three events have the same set of data (shown below) but are interpretted differently.
     See the class entry for details.
     """
-
-    name = 'HotkeyEvent'
-    is_player_action = True
-
     def __init__(self, frame, pid, data):
-        super(HotkeyEvent, self).__init__(frame, pid)
+        super(ControlGroupEvent, self).__init__(frame, pid)
 
         #: Index to the control group being modified
         self.control_group = data['control_group_index']
@@ -445,37 +408,32 @@ class HotkeyEvent(PlayerActionEvent):
         self.mask_data = data['remove_mask'][1]
 
 
-class SetToHotkeyEvent(HotkeyEvent):
+class SetControlGroupEvent(ControlGroupEvent):
     """
-    Extends :class:`HotkeyEvent`
+    Extends :class:`ControlGroupEvent`
 
     This event does a straight forward replace of the current control group contents
     with the player's current selection. This event doesn't have masks set.
     """
 
-    name = 'SetToHotkeyEvent'
 
-
-class AddToHotkeyEvent(HotkeyEvent):
+class AddToControlGroupEvent(SetControlGroupEvent):
     """
-    Extends :class:`HotkeyEvent`
+    Extends :class:`ControlGroupEvent`
 
     This event adds the current selection to the control group.
     """
 
-    name = 'AddToHotkeyEvent'
 
-
-class GetFromHotkeyEvent(HotkeyEvent):
+class GetControlGroupEvent(ControlGroupEvent):
     """
-    Extends :class:`HotkeyEvent`
+    Extends :class:`ControlGroupEvent`
+
     This event replaces the current selection with the contents of the control group.
     The mask data is used to limit that selection to units that are currently selectable.
     You might have 1 medivac and 8 marines on the control group but if the 8 marines are
     inside the medivac they cannot be part of your selection.
     """
-
-    name = 'GetFromHotkeyEvent'
 
 
 @loggable
@@ -485,9 +443,6 @@ class CameraEvent(GameEvent):
     It does not matter why the camera changed, this event simply records the current
     state of the camera after changing.
     """
-
-    name = 'CameraEvent'
-
     def __init__(self, frame, pid, data):
         super(CameraEvent, self).__init__(frame, pid)
 
@@ -515,8 +470,10 @@ class CameraEvent(GameEvent):
 
 @loggable
 class ResourceTradeEvent(GameEvent):
-    name = 'ResourceTradeEvent'
-
+    """
+    Generated when a player trades resources with another player. But not when fullfulling
+    resource requests.
+    """
     def __init__(self, frame, pid, data):
         super(ResourceTradeEvent, self).__init__(frame, pid)
 
@@ -552,8 +509,9 @@ class ResourceTradeEvent(GameEvent):
 
 
 class ResourceRequestEvent(GameEvent):
-    name = 'ResourceRequestEvent'
-
+    """
+    Generated when a player creates a resource request.
+    """
     def __init__(self, frame, pid, data):
         super(ResourceRequestEvent, self).__init__(frame, pid)
 
@@ -577,8 +535,9 @@ class ResourceRequestEvent(GameEvent):
 
 
 class ResourceRequestFulfillEvent(GameEvent):
-    name = 'ResourceRequestFulfillEvent'
-
+    """
+    Generated when a player accepts a resource request.
+    """
     def __init__(self, frame, pid, data):
         super(ResourceRequestFulfillEvent, self).__init__(frame, pid)
 
@@ -587,8 +546,9 @@ class ResourceRequestFulfillEvent(GameEvent):
 
 
 class ResourceRequestCancelEvent(GameEvent):
-    name = 'ResourceRequestCancelEvent'
-
+    """
+    Generated when a player cancels their resource request.
+    """
     def __init__(self, frame, pid, data):
         super(ResourceRequestCancelEvent, self).__init__(frame, pid)